+++ src/state/path_test.go
@@ -1,4 +1,4 @@
-package model
+package state
 
 import (
 	"fmt"
